requirements.txt
fastapi>=0.112
uvicorn[standard]>=0.30
pydantic>=2.7
python-dotenv>=1.0
requests>=2.32
SQLAlchemy>=2.0